+++ app/controllers/events_controller.rb
@@ -1,7 +1,6 @@
 class EventsController < ApplicationController
   before_action :find_user_character_ids, only: %i[index]
-  before_action :find_event_by_slug, only: %i[show]
-  before_action :find_event, only: %i[edit]
+  before_action :find_event_by_slug, only: %i[show edit]
 
   def index; end
 
@@ -11,7 +10,9 @@ class EventsController < ApplicationController
 
   def new; end
 
-  def edit; end
+  def edit
+    authorize! @event
+  end
 
   private
 
@@ -21,11 +22,6 @@ class EventsController < ApplicationController
 
   def find_event_by_slug
     @event = Event.find_by(slug: params[:id])
-    render_error('Object is not found') if @event.nil?
-  end
-
-  def find_event
-    @event = Event.find_by(id: params[:id])
     render_error('Object is not found') if @event.nil?
   end
 end