+++ spec/controllers/guilds_controller_spec.rb
@@ -46,4 +46,37 @@ RSpec.describe GuildsController, type: :controller do
       get :show, params: { locale: 'en', id: guild.slug }
     end
   end
+
+  describe 'GET#management' do
+    let!(:guild) { create :guild }
+
+    it_behaves_like 'User Auth'
+
+    context 'for logged user' do
+      sign_in_user
+
+      context 'for unexisted guild' do
+        it 'renders error template' do
+          get :management, params: { locale: 'en', id: 999 }
+
+          expect(response).to render_template 'shared/error'
+        end
+      end
+
+      context 'for existed guild' do
+        let!(:character) { create :character, guild: guild, world: guild.world, user: @current_user }
+        let!(:guild_role) { create :guild_role, guild: guild, character: character, name: 'gm' }
+
+        it 'renders management template' do
+          get :management, params: { locale: 'en', id: guild.slug }
+
+          expect(response).to render_template :management
+        end
+      end
+    end
+
+    def do_request
+      get :management, params: { locale: 'en', id: guild.slug }
+    end
+  end
 end